quick fix : getCircleArea parameter type change from mistaken int to suspected float

+++ set1/ex19.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <cmath>
 
-float getCircleArea(int D)
+float getCircleArea(float D)
 {
     const float PI = 3.141592653589793238;
